+++ contrib/opbot/botmd/botmd.go
@@ -10,10 +10,11 @@ import (
 
 // Bot represents the bot server.
 type Bot struct {
-	handler      metrics.Handler
-	server       *slacker.Slacker
-	cfg          config.Config
-	signozClient *signoz.Client
+	handler       metrics.Handler
+	server        *slacker.Slacker
+	cfg           config.Config
+	signozClient  *signoz.Client
+	signozEnabled bool
 }
 
 // NewBot creates a new bot server.
@@ -25,11 +26,22 @@ func NewBot(handler metrics.Handler, cfg config.Config) Bot {
 		server:  server,
 	}
 
-	bot.signozClient = signoz.NewClientFromUser(handler, cfg.SignozBaseURL, cfg.SignozEmail, cfg.SignozPassword)
+	// you should be able to run opbot even without signoz.
+	if cfg.SignozPassword != "" && cfg.SignozEmail != "" && cfg.SignozBaseURL != "" {
+		bot.signozClient = signoz.NewClientFromUser(handler, cfg.SignozBaseURL, cfg.SignozEmail, cfg.SignozPassword)
+		bot.signozEnabled = true
+	}
 
+	bot.addMiddleware(bot.tracingMiddleware(), bot.metricsMiddleware())
 	bot.addCommands(bot.traceCommand(), bot.rfqLookupCommand())
 
 	return bot
+}
+
+func (b *Bot) addMiddleware(middlewares ...slacker.CommandMiddlewareHandler) {
+	for _, middleware := range middlewares {
+		b.server.AddCommandMiddleware(middleware)
+	}
 }
 
 func (b *Bot) addCommands(commands ...*slacker.CommandDefinition) {