+++ contrib/screener-api/screener/screener.go
@@ -13,16 +13,18 @@ import (
 	"sync"
 	"time"
 
+	"github.com/synapsecns/sanguine/core/mapmutex"
+	"golang.org/x/sync/errgroup"
+
 	"github.com/gin-gonic/gin"
 	"github.com/gin-gonic/gin/binding"
 	"github.com/ipfs/go-log"
+	"github.com/synapsecns/sanguine/contrib/screener-api/chainalysis"
 	"github.com/synapsecns/sanguine/contrib/screener-api/client"
 	"github.com/synapsecns/sanguine/contrib/screener-api/config"
 	"github.com/synapsecns/sanguine/contrib/screener-api/db"
 	"github.com/synapsecns/sanguine/contrib/screener-api/db/sql"
 	"github.com/synapsecns/sanguine/contrib/screener-api/docs"
-	"github.com/synapsecns/sanguine/contrib/screener-api/screener/internal"
-	"github.com/synapsecns/sanguine/contrib/screener-api/trmlabs"
 	"github.com/synapsecns/sanguine/core"
 	"github.com/synapsecns/sanguine/core/dbcommon"
 	"github.com/synapsecns/sanguine/core/ginhelper"
@@ -30,10 +32,14 @@ import (
 	baseServer "github.com/synapsecns/sanguine/core/server"
 	"go.opentelemetry.io/otel/attribute"
 	"go.opentelemetry.io/otel/trace"
-	"golang.org/x/exp/slices"
 
 	swaggerfiles "github.com/swaggo/files"
 	ginSwagger "github.com/swaggo/gin-swagger"
+)
+
+const (
+	okResponse  = "OK"
+	errResponse = "ERROR"
 )
 
 // Screener is the interface for the screener.
@@ -42,16 +48,15 @@ type Screener interface {
 }
 
 type screenerImpl struct {
-	rulesManager internal.RulesetManager
-	thresholds   []config.VolumeThreshold
-	db           db.DB
-	router       *gin.Engine
-	metrics      metrics.Handler
-	cfg          config.Config
-	client       trmlabs.Client
-	blacklist    []string
-	blacklistMux sync.RWMutex
-	whitelist    []string
+	db                db.DB
+	router            *gin.Engine
+	metrics           metrics.Handler
+	cfg               config.Config
+	client            chainalysis.Client
+	whitelist         map[string]bool
+	blacklistCache    map[string]bool
+	blacklistCacheMux sync.RWMutex
+	requestMux        mapmutex.StringMapMutex
 }
 
 var logger = log.Logger("screener")
@@ -59,50 +64,84 @@ var logger = log.Logger("screener")
 // NewScreener creates a new screener.
 func NewScreener(ctx context.Context, cfg config.Config, metricHandler metrics.Handler) (_ Screener, err error) {
 	screener := screenerImpl{
-		metrics: metricHandler,
-		cfg:     cfg,
+		metrics:    metricHandler,
+		cfg:        cfg,
+		requestMux: mapmutex.NewStringMapMutex(),
 	}
 
 	docs.SwaggerInfo.Title = "Screener API"
 	docs.SwaggerInfo.Host = fmt.Sprintf("localhost:%d", cfg.Port)
 
-	screener.client, err = trmlabs.NewClient(cfg.TRMKey, core.GetEnv("TRM_URL", "https://api.trmlabs.com"))
-	if err != nil {
-		return nil, fmt.Errorf("could not create trm client: %w", err)
-	}
-	screener.thresholds = cfg.VolumeThresholds
+	screener.client = chainalysis.NewClient(metricHandler, cfg.RiskLevels, cfg.ChainalysisKey, core.GetEnv("CHAINALYSIS_URL", cfg.ChainalysisURL))
 
+	screener.blacklistCache = make(map[string]bool)
+	screener.whitelist = make(map[string]bool)
 	for _, item := range cfg.Whitelist {
-		screener.whitelist = append(screener.whitelist, strings.ToLower(item))
-	}
-
-	screener.rulesManager, err = setupScreener(cfg.Rulesets)
-	if err != nil {
-		return nil, fmt.Errorf("could not setup screener: %w", err)
+		screener.whitelist[strings.ToLower(item)] = true
 	}
 
 	dbType, err := dbcommon.DBTypeFromString(cfg.Database.Type)
 	if err != nil {
 		return nil, fmt.Errorf("could not get db type: %w", err)
 	}
-
 	screener.db, err = sql.Connect(ctx, dbType, cfg.Database.DSN, metricHandler)
 	if err != nil {
 		return nil, fmt.Errorf("could not connect to rules db: %w", err)
 	}
 
 	screener.router = ginhelper.New(logger)
-	screener.router.Use(screener.metrics.Gin())
-	screener.router.Handle(http.MethodGet, "/:ruleset/address/:address", screener.screenAddress)
+	screener.router.Use(screener.metrics.Gin()...)
 
-	screener.router.Handle(http.MethodPost, "/api/data/sync", ginhelper.TraceMiddleware(metricHandler.Tracer(), true), screener.authMiddleware(cfg), screener.blacklistAddress)
+	// Blacklist route
+	screener.router.POST("/api/data/sync", ginhelper.TraceMiddleware(metricHandler.Tracer(), true), screener.authMiddleware(cfg), screener.blacklistAddress)
+
+	// Screening routes
+	screener.router.GET("/address/:address", screener.screenAddress)
+	// deprecated and ruleset is not used, this is for backwards compatibility
+	screener.router.GET("/:ruleset/address/:address", screener.screenAddress)
+
+	// Swagger routes
 	screener.router.GET("/swagger/*any", ginSwagger.WrapHandler(swaggerfiles.Handler))
+
+	gin.SetMode(gin.ReleaseMode)
 
 	return &screener, nil
 }
 
+const blacklistScreenInterval = 15 * time.Second
+
+func (s *screenerImpl) Start(ctx context.Context) error {
+	// TODO: potential race condition here, if the blacklist is not fetched before the first request
+	// in practice chainalysis will catch
+	go func() {
+		// Fetch the blacklist at start.
+		s.fetchBlacklist(ctx)
+
+		for {
+			select {
+			case <-ctx.Done():
+				return
+			case <-time.After(blacklistScreenInterval):
+				s.fetchBlacklist(ctx)
+			}
+		}
+	}()
+
+	connection := baseServer.Server{}
+	err := connection.ListenAndServe(ctx, fmt.Sprintf(":%d", s.cfg.Port), s.router)
+	if err != nil {
+		return fmt.Errorf("could not start server: %w", err)
+	}
+	return nil
+}
+
 func (s *screenerImpl) fetchBlacklist(ctx context.Context) {
-	req, err := http.NewRequestWithContext(ctx, http.MethodGet, s.cfg.BlacklistURL, nil)
+	req, err := http.NewRequestWithContext(
+		ctx,
+		http.MethodGet,
+		s.cfg.BlacklistURL,
+		nil,
+	)
 	if err != nil {
 		logger.Errorf("could not create blacklist request: %s", err)
 		return
@@ -113,7 +152,6 @@ func (s *screenerImpl) fetchBlacklist(ctx context.Context) {
 		logger.Errorf("could not fetch blacklist: %s", err)
 		return
 	}
-
 	defer func() {
 		_ = resp.Body.Close()
 	}()
@@ -125,12 +163,94 @@ func (s *screenerImpl) fetchBlacklist(ctx context.Context) {
 		return
 	}
 
-	s.blacklistMux.Lock()
-	defer s.blacklistMux.Unlock()
-
+	s.blacklistCacheMux.Lock()
+	defer s.blacklistCacheMux.Unlock()
 	for _, item := range blacklist {
-		s.blacklist = append(s.blacklist, strings.ToLower(item))
+		s.blacklistCache[strings.ToLower(item)] = true
 	}
+}
+
+// screenAddress godoc
+// @Summary Screen an address for risk
+// @Description Screen an address using Chainalysis API to determine if it's high risk
+// @Tags Address Screening
+// @Accept  json
+// @Produce  json
+// @Param   address path string true "Address to be screened"
+// @Accept json
+// @Produce json
+// @Router /screen/{address} [get].
+func (s *screenerImpl) screenAddress(c *gin.Context) {
+	address := strings.ToLower(c.Param("address"))
+	if address == "" {
+		logger.Errorf("address is required")
+		c.JSON(http.StatusBadRequest, gin.H{"error": "address is required"})
+		return
+	}
+
+	// Check if the address is in the blacklist.
+	if s.isBlacklistedCache(address) {
+		c.JSON(http.StatusOK, gin.H{"risk": true})
+		return
+	}
+
+	// Check if the address is in the whitelist.
+	if _, whitelisted := s.whitelist[address]; whitelisted {
+		c.JSON(http.StatusOK, gin.H{"risk": false})
+		return
+	}
+
+	// prevent a single address from saturating the server.
+	// the only case this is useful is with a bad client that continuously sends requests for the same address.
+	// due to a goroutine leak, etc.
+	unlocker := s.requestMux.Lock(address)
+	defer unlocker.Unlock()
+
+	g, ctx := errgroup.WithContext(c.Request.Context())
+	var isAPIBlocked, isDBBlocked bool
+	g.Go(func() (err error) {
+		// If not, check db & Chainalysis for the risk assessment.
+		isAPIBlocked, err = s.client.ScreenAddress(ctx, address)
+		if err != nil {
+			return fmt.Errorf("error screening address: %w", err)
+		}
+		return nil
+	})
+
+	g.Go(func() (err error) {
+		isDBBlocked, err = s.isDBBlacklisted(ctx, address)
+		if err != nil {
+			return fmt.Errorf("error checking db: %w", err)
+		}
+		return nil
+	})
+
+	if err := g.Wait(); err != nil {
+		logger.Errorf("error screening address: %v", err)
+		c.JSON(http.StatusInternalServerError, gin.H{"error": err.Error()})
+		return
+	}
+
+	c.JSON(http.StatusOK, gin.H{"risk": isAPIBlocked || isDBBlocked})
+}
+
+func (s *screenerImpl) isDBBlacklisted(ctx context.Context, address string) (bool, error) {
+	_, err := s.db.GetBlacklistedAddress(ctx, address)
+	if err != nil && !errors.Is(err, db.ErrNoAddressNotFound) {
+		return false, fmt.Errorf("could not get blacklisted address: %w", err)
+	}
+
+	if errors.Is(err, db.ErrNoAddressNotFound) {
+		return false, nil
+	}
+
+	return true, nil
+}
+
+func (s *screenerImpl) isBlacklistedCache(address string) bool {
+	s.blacklistCacheMux.RLock()
+	defer s.blacklistCacheMux.RUnlock()
+	return s.blacklistCache[address]
 }
 
 // @dev Protected Method
@@ -148,73 +268,74 @@ func (s *screenerImpl) fetchBlacklist(ctx context.Context) {
 func (s *screenerImpl) blacklistAddress(c *gin.Context) {
 	var err error
 	ctx, span := s.metrics.Tracer().Start(c.Request.Context(), "blacklistAddress")
-	defer func() {
-		metrics.EndSpanWithErr(span, err)
-	}()
+	defer metrics.EndSpanWithErr(span, err)
 
 	var blacklistBody client.BlackListBody
 
 	// Grab the body of the JSON request and unmarshal it into the blacklistBody struct.
 	if err := c.ShouldBindBodyWith(&blacklistBody, binding.JSON); err != nil {
-		c.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
+		c.JSON(http.StatusBadRequest, gin.H{"error": errResponse})
 		return
 	}
 
-	span.SetAttributes(attribute.String("type", blacklistBody.Type))
-	span.SetAttributes(attribute.String("id", blacklistBody.ID))
-	span.SetAttributes(attribute.String("data", blacklistBody.Data))
-	span.SetAttributes(attribute.String("network", blacklistBody.Network))
-	span.SetAttributes(attribute.String("tag", blacklistBody.Tag))
-	span.SetAttributes(attribute.String("remark", blacklistBody.Remark))
-	span.SetAttributes(attribute.String("address", blacklistBody.Address))
+	span.SetAttributes(
+		attribute.String("type", blacklistBody.Type),
+		(attribute.String("id", blacklistBody.ID)),
+		(attribute.String("network", blacklistBody.Data.Network)),
+		(attribute.String("tag", blacklistBody.Data.Tag)),
+		(attribute.String("remark", blacklistBody.Data.Remark)),
+		(attribute.String("address", blacklistBody.Data.Address)),
+	)
 
 	blacklistedAddress := db.BlacklistedAddress{
 		Type:    blacklistBody.Type,
 		ID:      blacklistBody.ID,
-		Data:    blacklistBody.Data,
-		Network: blacklistBody.Network,
-		Tag:     blacklistBody.Tag,
-		Remark:  blacklistBody.Remark,
-		Address: strings.ToLower(blacklistBody.Address),
+		Network: blacklistBody.Data.Network,
+		Tag:     blacklistBody.Data.Tag,
+		Remark:  blacklistBody.Data.Remark,
+		Address: strings.ToLower(blacklistBody.Data.Address),
 	}
+
+	s.blacklistCacheMux.Lock()
+	defer s.blacklistCacheMux.Unlock()
+	s.blacklistCache[blacklistBody.Data.Address] = true
 
 	switch blacklistBody.Type {
 	case "create":
 		if err := s.db.PutBlacklistedAddress(ctx, blacklistedAddress); err != nil {
 			span.AddEvent("error", trace.WithAttributes(attribute.String("error", err.Error())))
-			c.JSON(http.StatusInternalServerError, gin.H{"error": err.Error()})
+			c.JSON(http.StatusInternalServerError, gin.H{"error": errResponse})
 			return
 		}
 
-		span.AddEvent("blacklistedAddress", trace.WithAttributes(attribute.String("address", blacklistBody.Address)))
-		c.JSON(http.StatusOK, gin.H{"status": "success"})
+		span.AddEvent("blacklistedAddress", trace.WithAttributes(attribute.String("address", blacklistBody.Data.Address)))
+		c.JSON(http.StatusOK, gin.H{"status": okResponse})
 		return
 
 	case "update":
 		if err := s.db.UpdateBlacklistedAddress(ctx, blacklistedAddress.ID, blacklistedAddress); err != nil {
 			span.AddEvent("error", trace.WithAttributes(attribute.String("error", err.Error())))
-			c.JSON(http.StatusInternalServerError, gin.H{"error": err.Error()})
+			c.JSON(http.StatusInternalServerError, gin.H{"error": errResponse})
 			return
 		}
 
-		span.AddEvent("blacklistedAddress", trace.WithAttributes(attribute.String("address", blacklistBody.Address)))
-		c.JSON(http.StatusOK, gin.H{"status": "success"})
+		span.AddEvent("blacklistedAddress", trace.WithAttributes(attribute.String("address", blacklistBody.Data.Address)))
+		c.JSON(http.StatusOK, gin.H{"status": okResponse})
 		return
 
 	case "delete":
 		if err := s.db.DeleteBlacklistedAddress(ctx, blacklistedAddress.ID); err != nil {
 			span.AddEvent("error", trace.WithAttributes(attribute.String("error", err.Error())))
-			c.JSON(http.StatusInternalServerError, gin.H{"error": err.Error()})
+			c.JSON(http.StatusInternalServerError, gin.H{"error": errResponse})
 			return
 		}
 
-		span.AddEvent("blacklistedAddress", trace.WithAttributes(attribute.String("address", blacklistBody.Address)))
-		c.JSON(http.StatusOK, gin.H{"status": "success"})
+		span.AddEvent("blacklistedAddress", trace.WithAttributes(attribute.String("address", blacklistBody.Data.Address)))
+		c.JSON(http.StatusOK, gin.H{"status": okResponse})
 		return
 
 	default:
-		span.AddEvent("error", trace.WithAttributes(attribute.String("error", err.Error())))
-		c.JSON(http.StatusBadRequest, gin.H{"error": "invalid type"})
+		c.JSON(http.StatusBadRequest, gin.H{"error": errResponse})
 		return
 	}
 }
@@ -233,17 +354,34 @@ func (s *screenerImpl) authMiddleware(cfg config.Config) gin.HandlerFunc {
 		signature := c.Request.Header.Get("X-Signature-signature")
 		queryString := c.Request.URL.RawQuery
 
-		bodyBytes, err := io.ReadAll(c.Request.Body)
+		bodyBz, err := io.ReadAll(c.Request.Body)
 		if err != nil {
-			c.JSON(http.StatusBadRequest, gin.H{"error": "could not read request body"})
+			c.JSON(http.StatusBadRequest, gin.H{"error": errResponse})
 			c.Abort()
 			return
 		}
-		c.Request.Body = io.NopCloser(bytes.NewBuffer(bodyBytes))
-		bodyStr := string(bodyBytes)
+		// Put it back so we can read it again for DB operations.
+		c.Request.Body = io.NopCloser(bytes.NewBuffer(bodyBz))
 
-		message := fmt.Sprintf("%s%s%s%s%s%s%s",
-			appID, timestamp, nonce, "POST", "/api/data/sync", queryString, bodyStr)
+		bodyStr, err := core.BytesToJSONString(bodyBz)
+		if err != nil {
+			c.JSON(http.StatusBadRequest, gin.H{"error": errResponse})
+			c.Abort()
+			return
+		}
+
+		var message string
+		if len(queryString) > 0 {
+			message = fmt.Sprintf(
+				"%s;%s;%s;%s;%s;%s;%s",
+				appID, timestamp, nonce, "POST", "/api/data/sync", queryString, bodyStr,
+			)
+		} else {
+			message = fmt.Sprintf(
+				"%s;%s;%s;%s;%s;%s",
+				appID, timestamp, nonce, "POST", "/api/data/sync", bodyStr,
+			)
+		}
 
 		expectedSignature := client.GenerateSignature(cfg.AppSecret, message)
 
@@ -263,132 +401,11 @@ func (s *screenerImpl) authMiddleware(cfg config.Config) gin.HandlerFunc {
 				"error",
 				trace.WithAttributes(attribute.String("error", "Invalid signature"+expectedSignature)),
 			)
-			c.JSON(http.StatusUnauthorized, gin.H{"error": "Invalid signature"})
+			c.JSON(http.StatusUnauthorized, gin.H{"error": errResponse})
 			c.Abort()
 			return
 		}
 		span.AddEvent("success", trace.WithAttributes(attribute.String("message", "Valid signature"+expectedSignature)))
 		c.Next()
 	}
-}
-
-func (s *screenerImpl) Start(ctx context.Context) error {
-	// TODO: potential race condition here, if the blacklist is not fetched before the first request
-	// in practice trm will catch
-	go func() {
-		for {
-			if s.cfg.BlacklistURL != "" {
-				s.fetchBlacklist(ctx)
-				time.Sleep(1 * time.Second * 15)
-			}
-		}
-	}()
-	connection := baseServer.Server{}
-	err := connection.ListenAndServe(ctx, fmt.Sprintf(":%d", s.cfg.Port), s.router)
-	if err != nil {
-		return fmt.Errorf("could not start server: %w", err)
-	}
-	return nil
-}
-
-// screenAddress returns whether an address is risky or not given a ruleset.
-// @Summary Screen address for risk
-// @Description Assess the risk associated with a given address using specified rulesets.
-// @Tags address
-// @Accept  json
-// @Produce  json
-// @Param ruleset query string true "Ruleset to use for screening the address"
-// @Param address query string true "Address to be screened"
-// @Success 200 {object} map[string]bool "Returns the risk assessment result"
-// @Failure 400 {object} map[string]string "Returns error if the required parameters are missing or invalid"
-// @Failure 500 {object} map[string]string "Returns error if there are problems processing the indicators"
-// @Router /screen/{ruleset}/{address} [get].
-func (s *screenerImpl) screenAddress(c *gin.Context) {
-	var err error
-
-	ruleset := strings.ToLower(c.Param("ruleset"))
-	if ruleset == "" {
-		c.JSON(http.StatusBadRequest, gin.H{"error": "ruleset is required"})
-		return
-	}
-
-	address := strings.ToLower(c.Param("address"))
-	if address == "" {
-		c.JSON(http.StatusBadRequest, gin.H{"error": "address is required"})
-		return
-	}
-
-	s.blacklistMux.RLock()
-	if slices.Contains(s.blacklist, address) {
-		c.JSON(http.StatusOK, gin.H{"risk": true})
-		s.blacklistMux.RUnlock()
-		return
-	}
-	s.blacklistMux.RUnlock()
-
-	if slices.Contains(s.whitelist, address) {
-		c.JSON(http.StatusOK, gin.H{"risk": false})
-		return
-	}
-
-	ctx, span := s.metrics.Tracer().Start(c.Request.Context(), "screenAddress", trace.WithAttributes(attribute.String("address", address)))
-	defer func() {
-		metrics.EndSpanWithErr(span, err)
-	}()
-
-	currentRules := s.rulesManager.GetRuleset(ruleset)
-	if currentRules == nil {
-		c.JSON(http.StatusBadRequest, gin.H{"error": "ruleset not found"})
-		return
-	}
-
-	goodUntil := time.Now().Add(-1 * s.cfg.GetCacheTime(ruleset))
-	var indicators []trmlabs.AddressRiskIndicator
-	if indicators, err = s.getIndicators(ctx, address, goodUntil); err != nil {
-		c.JSON(http.StatusInternalServerError, gin.H{"error": err.Error()})
-		return
-	}
-
-	var hasIndicator bool
-	if hasIndicator, err = currentRules.HasAddressIndicators(s.thresholds, indicators...); err != nil {
-		c.JSON(http.StatusOK, gin.H{"risk": true})
-		return
-	}
-
-	c.JSON(http.StatusOK, gin.H{"risk": hasIndicator})
-}
-
-func (s *screenerImpl) getIndicators(parentCtx context.Context, address string, goodUntil time.Time) (indicators []trmlabs.AddressRiskIndicator, err error) {
-	ctx, span := s.metrics.Tracer().Start(parentCtx, "get-indicators")
-	defer func() {
-		// nolint: errchkjson
-		marshalledIndicators, _ := json.Marshal(indicators)
-		span.AddEvent("indicators", trace.WithAttributes(attribute.String("indicators", string(marshalledIndicators))))
-		metrics.EndSpanWithErr(span, err)
-	}()
-
-	riskIndicators, err := s.db.GetAddressIndicators(ctx, address, goodUntil)
-	if err == nil {
-		return riskIndicators, nil
-	}
-
-	if !errors.Is(err, db.ErrNoAddressNotCached) {
-		return nil, fmt.Errorf("could not get address indicators: %w", err)
-	}
-
-	response, err := s.client.ScreenAddress(ctx, address)
-	if err != nil {
-		return nil, fmt.Errorf("could not screen address: %w", err)
-	}
-
-	for _, ri := range response {
-		riskIndicators = append(riskIndicators, ri.AddressRiskIndicators...)
-	}
-
-	err = s.db.PutAddressIndicators(ctx, address, riskIndicators)
-	if err != nil {
-		return nil, fmt.Errorf("could not put address indicators: %w", err)
-	}
-
-	return riskIndicators, nil
 }