+++ packages/explorer-ui/components/TransactionTable/TableBody.tsx
@@ -7,9 +7,9 @@ export function TableBody({ rows }) {
         <TableRow
           key={index}
           items={row.items}
-          className={` hover:border-y border-slate-800 hover:bg-slate-900 hover:cursor-pointer ${
-            index === 0 ? 'border-t-0' : ''} ${index === rows.length - 1 ? 'border-b-0' : ''}`
-          }
+          className={`hover:bg-[#0F172A] hover:cursor-pointer ${
+            index % 2 === 0 ? 'bg-[#101018]' : 'bg-[#100C13]'
+          } ${index === 0 ? 'border-t-0' : ''} ${index === rows.length - 1 ? 'border-b-0' : ''}`}
         />
       ))}
     </tbody>