+++ packages/synapse-interface/components/Maintenance/components/MaintenanceBanner.tsx
@@ -3,16 +3,6 @@ import { getCountdownTimeStatus } from './EventCountdownProgressBar'
 import { AnnouncementBanner } from './AnnouncementBanner'
 import { isNull } from 'lodash'
 
-/**
- * Handles constructing and updating instances of Announcement Banners based on specific start / end time.
- * If end date is null, indefinitely display banner.
- *
- * @param id Unique ID to prevent conflicts with other banner instances.
- *           Assign ID $MMDDYYYY-$BANNER_NAME format (e.g 03132024-ETH-DENCUN)
- * @param startDate Start time to display banner
- * @param endDate End time to remove banner
- * @param bannerMessage - Message to display in banner
- */
 export const MaintenanceBanner = ({
   id,
   startDate,
@@ -38,7 +28,7 @@ export const MaintenanceBanner = ({
     return (
       <AnnouncementBanner
         bannerId={id}
-        bannerContents={bannerMessage}
+        bannerContent={bannerMessage}
         startDate={startDate}
         endDate={endDate}
       />