+++ packages/explorer-ui/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.1.46](https://github.com/synapsecns/sanguine/compare/@synapsecns/explorer-ui@0.1.45...@synapsecns/explorer-ui@0.1.46) (2024-06-28)
+
+**Note:** Version bump only for package @synapsecns/explorer-ui
+
+
+
+
+
 ## [0.1.45](https://github.com/synapsecns/sanguine/compare/@synapsecns/explorer-ui@0.1.44...@synapsecns/explorer-ui@0.1.45) (2024-06-16)
 
 **Note:** Version bump only for package @synapsecns/explorer-ui